+++ android/src/main/kotlin/com/example/flutter_local_network_ios_plus/FlutterLocalNetworkIosPlusPlugin.kt
@@ -1,4 +1,4 @@
-package com.example.flutter_local_network_ios
+package com.example.flutter_local_network_ios_plus
 
 import androidx.annotation.NonNull
 
@@ -8,8 +8,8 @@ import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 
-/** FlutterLocalNetworkIosPlugin */
-class FlutterLocalNetworkIosPlugin: FlutterPlugin, MethodCallHandler {
+/** FlutterLocalNetworkIosPlusPlugin */
+class FlutterLocalNetworkIosPlusPlugin: FlutterPlugin, MethodCallHandler {
   /// The MethodChannel that will the communication between Flutter and native Android
   ///
   /// This local reference serves to register the plugin with the Flutter Engine and unregister it
@@ -17,7 +17,7 @@ class FlutterLocalNetworkIosPlugin: FlutterPlugin, MethodCallHandler {
   private lateinit var channel : MethodChannel
 
   override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
-    channel = MethodChannel(flutterPluginBinding.binaryMessenger, "flutter_local_network_ios")
+    channel = MethodChannel(flutterPluginBinding.binaryMessenger, "flutter_local_network_ios_plus")
     channel.setMethodCallHandler(this)
   }