+++ lib/flutter_local_network_ios_plus.dart
@@ -1,0 +1,11 @@
+import 'flutter_local_network_ios_plus_platform_interface.dart';
+
+class FlutterLocalNetworkIosPlus {
+  Future<String?> getPlatformVersion() {
+    return FlutterLocalNetworkIosPlusPlatform.instance.getPlatformVersion();
+  }
+
+  Future<bool?> requestAuthorization() {
+    return FlutterLocalNetworkIosPlusPlatform.instance.requestAuthorization();
+  }
+}