+++ test/flutter_local_network_ios_plus_method_channel_test.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/services.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:flutter_local_network_ios_plus/flutter_local_network_ios_plus_method_channel.dart';
+
+void main() {
+  MethodChannelFlutterLocalNetworkIosPlus platform =
+      MethodChannelFlutterLocalNetworkIosPlus();
+  const MethodChannel channel = MethodChannel('flutter_local_network_ios_plus');
+
+  TestWidgetsFlutterBinding.ensureInitialized();
+
+  setUp(() {
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(
+      channel,
+      (MethodCall methodCall) async {
+        return '42';
+      },
+    );
+  });
+
+  tearDown(() {
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+        .setMockMethodCallHandler(
+      channel,
+      null,
+    );
+  });
+
+  test('getPlatformVersion', () async {
+    expect(await platform.getPlatformVersion(), '42');
+  });
+}